requirements.txt
beautifulsoup4==4.12.3
numpy==1.26.4
Pillow==10.2.0
pygraphviz==1.12
python_Levenshtein==0.25.0
scipy==1.12.0
selenium==4.17.2
tqdm==4.66.2
